requirements.txt
fastapi
uvicorn
motor
pandas
python-dotenv
pydantic
requests
